Make configuration file unnecessary

+++ standalone/bin/ljb-setup-start.sh
@@ -2,5 +2,8 @@
 
 . "$SCRIPT_DIR/bin/ljb-setup-conf.sh"
 
-rm -rf "$LIFERAY_HOME"
-mkdir -p "$LIFERAY_HOME"
+if [ "x$LIFERAY_HOME" != "x" ]
+then
+	rm -rf "$LIFERAY_HOME"
+	mkdir -p "$LIFERAY_HOME"
+fi
